Incremental Lock as BigUint
