+++ src/lib.rs
@@ -13,7 +13,8 @@ include!("crates.rs");
 
 pub mod script;
 #[allow(dead_code)]
-mod logicalstamp;
+mod timestamp;
+mod incrementalstamp;
 #[allow(dead_code)]
 pub mod pubsub;
 pub mod storage;
