+++ src/main.rs
@@ -16,7 +16,8 @@ extern crate log4rs;
 
 pub mod script;
 pub mod server;
-pub mod logicalstamp;
+pub mod timestamp;
+pub mod incrementalstamp;
 pub mod pubsub;
 pub mod storage;
