+++ src/script/mod_incremental_lock.rs
@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //!
-//! # Atomic Counter Value: ACV
+//! # Atomic Counter Value: ACV, an incremental stamp 
 //!
 //! This module handles acces to a global usize count for increasing count needs
 //!
@@ -16,8 +16,7 @@ word!(ACV, b"\x83ACV");
 
 use std::marker::PhantomData;
 use super::{Module, PassResult, Error, Env, EnvId};
-use logicalstamp;
-use byteorder::{ByteOrder, BigEndian};
+use incrementalstamp;
 
 pub struct Handler<'a> {
     phantom: PhantomData<&'a ()>
@@ -38,24 +37,11 @@ impl<'a> Handler<'a> {
     #[inline]
     pub fn handle_acv(&mut self, env: &mut Env<'a>, word: &'a [u8], _: EnvId) -> PassResult<'a> {
         if word == ACV {
-            let c = logicalstamp::acv_count();
-
-            if cfg!(target_pointer_width = "32"){
-                let mut buf = [0; 4];
-                let val = c as u32;
-                BigEndian::write_u32(&mut buf, val);
-                let slice = alloc_and_write!(&buf, env);
-                env.push(slice);
-            }
-
-            if cfg!(target_pointer_width = "64") {
-                let mut buf = [0; 8];
-                let val = c as u64;
-                BigEndian::write_u64(&mut buf, val);
-                let slice = alloc_and_write!(&buf, env);
-                env.push(slice);
-            }
-            //env.push(slice);
+            let c = incrementalstamp::count();
+            let buf = c.to_bytes_be();
+            let slice = alloc_and_write!(&buf, env);
+            env.push(slice);
+            
             Ok(())
         }else {
             //Does not match word
